% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionary.R
\name{ext_binary}
\alias{ext_binary}
\title{Common file extensions (binary)}
\usage{
ext_binary(flat = FALSE)
}
\arguments{
\item{flat}{Flatten the list and return a vector?}
}
\value{
A list or vector of standard binary file extensions.
}
\description{
Common file extensions (binary)
}
\section{Specification}{

\if{latex}{
  \itemize{
    \item Return a named list of common binary file extensions in R packages.
  }
}
\if{html}{The contents of this section are shown in PDF user manual only.}
}

\examples{
ext_binary()
}
