% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionary.R
\name{ext_text}
\alias{ext_text}
\title{Common file extensions (text)}
\usage{
ext_text(flat = FALSE)
}
\arguments{
\item{flat}{Flatten the list and return a vector?}
}
\value{
A list or vector of standard text file extensions.
}
\description{
Common file extensions (text)
}
\section{Specification}{

\if{latex}{
  \itemize{
    \item Return a named list of common text file extensions in R packages.
  }
}
\if{html}{The contents of this section are shown in PDF user manual only.}
}

\examples{
ext_text()
}
