% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{file_default}
\alias{file_default}
\title{File specification (default combination)}
\usage{
file_default()
}
\value{
A list of file specifications.
}
\description{
A default combination of common file specifications.
}
\section{Specification}{

\if{latex}{
  \itemize{
    \item Use file specification template functions to generate and
    return a list of file specifications that cover a default
    set of the common naming patterns of files in source R packages.
  }
}
\if{html}{The contents of this section are shown in PDF user manual only.}
}

\examples{
file_default()
}
