% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{file_ectd}
\alias{file_ectd}
\title{File specification (eCTD submission)}
\usage{
file_ectd()
}
\value{
A list of file specifications.
}
\description{
A combination of file specifications for eCTD submissions.
}
\section{Specification}{

\if{latex}{
  \itemize{
    \item Use file specification template functions to generate and
    return a list of file specifications that cover the a set of the
    common naming patterns of files in source R packages for
    eCTD submissions.
  }
}
\if{html}{The contents of this section are shown in PDF user manual only.}
}

\examples{
file_ectd()
}
