% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionary.R
\name{file_name_patterns}
\alias{file_name_patterns}
\alias{pattern_file_root_core}
\alias{pattern_file_root_all}
\alias{pattern_file_sanitize}
\title{Common File name patterns}
\usage{
pattern_file_root_core()

pattern_file_root_all()

pattern_file_sanitize()
}
\value{
A vector of file name patterns.
}
\description{
Common File name patterns
}
\section{Specification}{

\if{latex}{
  \itemize{
    \item Return a vector of filename patterns for matching files.
  }
}
\if{html}{The contents of this section are shown in PDF user manual only.}
}

