% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collection.R
\name{is_file_collection}
\alias{is_file_collection}
\title{Is this a file collection object?}
\usage{
is_file_collection(object)
}
\arguments{
\item{object}{Any R object.}
}
\value{
Logical. \code{TRUE} if it is a file collection object,
\code{FALSE} otherwise.
}
\description{
Is this a file collection object?
}
\section{Specification}{

\if{latex}{
  \itemize{
    \item Check if the object has the class \code{file_collection}.
  }
}
\if{html}{The contents of this section are shown in PDF user manual only.}
}

\examples{
system.file("examples/pkg1/", package = "pkglite") \%>\%
  collate(file_default()) \%>\%
  is_file_collection()
}
