% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec.R
\name{is_file_spec}
\alias{is_file_spec}
\title{Is this a file specification object?}
\usage{
is_file_spec(object)
}
\arguments{
\item{object}{Any R object}
}
\value{
Logical. \code{TRUE} if it is a file specification object,
\code{FALSE} otherwise.
}
\description{
Is this a file specification object?
}
\section{Specification}{

\if{latex}{
  \itemize{
    \item Check if the input object class contains \code{file_spec}.
  }
}
\if{html}{The contents of this section are shown in PDF user manual only.}
}

\examples{
file_spec(
  "R/",
  pattern = "\\\\.R$", format = "text",
  recursive = FALSE, ignore_case = TRUE, all_files = FALSE
) \%>\%
  is_file_spec()
}
