% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{\%>\%}
\alias{\%>\%}
\title{Pipe operator}
\usage{
lhs \%>\% rhs
}
\value{
The evaluation result of \code{rhs(lhs)},
if \code{rhs} is a function and \code{lhs} is an object.
See \link[magrittr:pipe]{magrittr::\%>\%} for more details.
}
\description{
See \code{magrittr::\link[magrittr:pipe]{\%>\%}} for details.
}
\keyword{internal}
