% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collection.R
\name{print.file_collection}
\alias{print.file_collection}
\title{Print a file collection}
\usage{
\method{print}{file_collection}(x, ...)
}
\arguments{
\item{x}{An object of class \code{file_collection}.}

\item{...}{Additional parameters for \code{\link[=print]{print()}} (not used).}
}
\value{
The input \code{file_collection} object.
}
\description{
Print a file collection
}
\section{Specification}{

\if{latex}{
  \itemize{
    \item Print the metadata and the data frame in a file collection object.
  }
}
\if{html}{The contents of this section are shown in PDF user manual only.}
}

\examples{
fc <- system.file("examples/pkg1/", package = "pkglite") \%>\%
  collate(file_default())
fc
}
