% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec.R
\name{print.file_spec}
\alias{print.file_spec}
\title{Print a file specification}
\usage{
\method{print}{file_spec}(x, ...)
}
\arguments{
\item{x}{An object of class \code{file_spec}.}

\item{...}{Additional parameters for \code{\link[=print]{print()}} (not used).}
}
\value{
The input \code{file_spec} object.
}
\description{
Print a file specification
}
\section{Specification}{

\if{latex}{
  \itemize{
    \item Print the elements in the file specification object with cli.
  }
}
\if{html}{The contents of this section are shown in PDF user manual only.}
}

\examples{
fs <- file_spec(
  "R/",
  pattern = "\\\\.R$", format = "text",
  recursive = FALSE, ignore_case = TRUE, all_files = FALSE
)
fs
}
