% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collection.R
\name{sanitize}
\alias{sanitize}
\alias{sanitize.file_collection}
\title{Sanitize file collection}
\usage{
sanitize(x)

\method{sanitize}{file_collection}(x)
}
\arguments{
\item{x}{File collection.}
}
\value{
Sanitized file collection.
}
\description{
Remove commonly excluded files from a file collection.
}
\section{Specification}{

\if{latex}{
  \itemize{
    \item Remove the files whose names match certain patterns from the
    file collection and return a sanitized file collection.
  }
}
\if{html}{The contents of this section are shown in PDF user manual only.}
}

\examples{
system.file("examples/pkg1/", package = "pkglite") \%>\%
  collate(file_default()) \%>\%
  sanitize()
}
