% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collate.R
\name{sanitize_file_collection}
\alias{sanitize_file_collection}
\title{Sanitize file collection}
\usage{
sanitize_file_collection(x)
}
\arguments{
\item{x}{a file collection object}
}
\value{
The sanitized file collection object.
}
\description{
Remove commonly excluded files and directories from a file collection.
}
\section{Specification}{

\if{latex}{
  \itemize{
    \item Remove the files whose names match certain patterns from the
    file collection and return a sanitized file collection.
  }
}
\if{html}{The contents of this section are shown in PDF user manual only.}
}

\examples{
system.file("examples/pkg1/", package = "pkglite") \%>\%
  collate(file_default()) \%>\%
  sanitize_file_collection()
}
