% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkglite-package.R
\docType{package}
\name{pkglite-package}
\alias{pkglite}
\alias{pkglite-package}
\title{pkglite: Compact Package Representations}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A tool, grammar, and standard to represent and exchange R package source code as text files. Converts one or more source packages to a text file and restores the package structures from the file.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://merck.github.io/pkglite/}
  \item \url{https://github.com/Merck/pkglite}
  \item Report bugs at \url{https://github.com/Merck/pkglite/issues}
}

}
\author{
\strong{Maintainer}: Nan Xiao \email{nan.xiao1@merck.com} (\href{https://orcid.org/0000-0002-0250-5673}{ORCID})

Authors:
\itemize{
  \item Yilong Zhang \email{yilong.zhang@merck.com}
  \item Keaven Anderson \email{keaven_anderson@merck.com}
}

Other contributors:
\itemize{
  \item Amin Shirazi \email{amin.shirazi@merck.com} [contributor]
  \item Jeff Cheng \email{jeff.cheng@merck.com} [contributor]
  \item Merck Sharp & Dohme Corp [copyright holder]
}

}
\keyword{internal}
